+++ packages/video_player_avplay/tizen/src/plus_player.cc
@@ -657,12 +657,98 @@ void PlusPlayer::OnResourceConflicted(void *user_data) {
   self->SendError("PlusPlayer error", "Resource conflicted");
 }
 
+std::string GetErrorMessage(plusplayer::ErrorType error_code) {
+  switch (error_code) {
+    case plusplayer::ErrorType::kNone: {
+      return "Successful";
+    }
+    case plusplayer::ErrorType::kOutOfMemory: {
+      return "Out of memory";
+    }
+    case plusplayer::ErrorType::kInvalidParameter: {
+      return "Invalid parameter";
+    }
+    case plusplayer::ErrorType::kNoSuchFile: {
+      return "No such file or directory";
+    }
+    case plusplayer::ErrorType::kInvalidOperation: {
+      return "Invalid operation";
+    }
+    case plusplayer::ErrorType::kFileNoSpaceOnDevice: {
+      return "No space left on the device";
+    }
+    case plusplayer::ErrorType::kFeatureNotSupportedOnDevice: {
+      return "Not supported file on this device";
+    }
+    case plusplayer::ErrorType::kSeekFailed: {
+      return "Seek operation failure";
+    }
+    case plusplayer::ErrorType::kInvalidState: {
+      return "Invalid player state";
+    }
+    case plusplayer::ErrorType::kNotSupportedFile: {
+      return "File format not supported";
+    }
+    case plusplayer::ErrorType::kNotSupportedFormat: {
+      return "Not supported media format";
+    }
+    case plusplayer::ErrorType::kInvalidUri: {
+      return "Invalid URI";
+    }
+    case plusplayer::ErrorType::kSoundPolicy: {
+      return "Sound policy error";
+    }
+    case plusplayer::ErrorType::kConnectionFailed: {
+      return "Streaming connection failed";
+    }
+    case plusplayer::ErrorType::kVideoCaptureFailed: {
+      return "Video capture failed";
+    }
+    case plusplayer::ErrorType::kDrmExpired: {
+      return "DRM license expired";
+    }
+    case plusplayer::ErrorType::kDrmNoLicense: {
+      return "DRM no license";
+    }
+    case plusplayer::ErrorType::kDrmFutureUse: {
+      return "License for future use";
+    }
+    case plusplayer::ErrorType::kDrmNotPermitted: {
+      return "DRM format not permitted";
+    }
+    case plusplayer::ErrorType::kResourceLimit: {
+      return "Resource limit";
+    }
+    case plusplayer::ErrorType::kPermissionDenied: {
+      return "Permission denied";
+    }
+    case plusplayer::ErrorType::kServiceDisconnected: {
+      return "Service disconnected";
+    }
+    case plusplayer::ErrorType::kBufferSpace: {
+      return "No buffer space available";
+    }
+    case plusplayer::ErrorType::kNotSupportedAudioCodec: {
+      return "Not supported audio codec but video can be played";
+    }
+    case plusplayer::ErrorType::kNotSupportedVideoCodec: {
+      return "Not supported video codec but audio can be played";
+    }
+    case plusplayer::ErrorType::kNotSupportedSubtitle: {
+      return "Not supported subtitle format";
+    }
+    default:
+      return "Not defined error";
+  }
+}
+
 void PlusPlayer::OnError(const plusplayer::ErrorType &error_code,
                          void *user_data) {
   LOG_ERROR("[PlusPlayer] Error code: %d", error_code);
   PlusPlayer *self = reinterpret_cast<PlusPlayer *>(user_data);
 
-  self->SendError("[PlusPlayer] OnError", "");
+  self->SendError("[PlusPlayer] error",
+                  std::string("Error: ") + GetErrorMessage(error_code));
 }
 
 void PlusPlayer::OnErrorMsg(const plusplayer::ErrorType &error_code,
@@ -670,7 +756,7 @@ void PlusPlayer::OnErrorMsg(const plusplayer::ErrorType &error_code,
   LOG_ERROR("[PlusPlayer] Error code: %d, message: %s.", error_code, error_msg);
   PlusPlayer *self = reinterpret_cast<PlusPlayer *>(user_data);
 
-  self->SendError("PlusPlayer error", error_msg);
+  self->SendError("PlusPlayer error", std::string("Error: ") + error_msg);
 }
 
 void PlusPlayer::OnDrmInitData(int *drm_handle, unsigned int len,